python/scratch/requirements.txt
azure-core==1.32.0
azure-cosmos==4.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
